Added spatial allocation
Spatial allocation works. Functions to build regions from shapefiles
are 99 percent of the way there. Tests with census shapefiles creates
a weird double counting error between north and south carolina. Not sure
if this is my method of the resolution of the data. Either way, with
correctly built region files, the approach works fine.