Improving pyshp to shapely conversion
pyshp returns all points as a list with segments starts identified in the parts list. This commit identifies the outer shell by assuming it has the most points. Fort the test data (cb_2018_us_state_500k), this assumption is great.

The easiest approach that does not require this assumption would be to move to something like fiona or gdal. Both of these are harder for users to install.

At the moment, I am prioritizing ease of installation over perfection. Depending on the problems in the future, that might change.